feat(image-uploader): rename image upload component with preview and removal functionality
